refactor: extract lookup into core module
