+++ src/core.rs
@@ -1,0 +1,67 @@
+use crate::database::DatastoreStorage;
+use crate::google::datastore::v1::{
+    Entity, EntityResult, LookupRequest, LookupResponse,
+};
+use std::collections::HashMap;
+use std::sync::Arc;
+use std::time::SystemTime;
+use tokio::sync::RwLock;
+use tonic::Status;
+
+pub(crate) fn system_time_to_timestamp(time: SystemTime) -> pbjson_types::Timestamp {
+    match time.duration_since(SystemTime::UNIX_EPOCH) {
+        Ok(duration) => pbjson_types::Timestamp {
+            seconds: duration.as_secs() as i64,
+            nanos: duration.subsec_nanos() as i32,
+        },
+        Err(_) => pbjson_types::Timestamp::default(),
+    }
+}
+
+pub async fn lookup(
+    storage: &Arc<RwLock<DatastoreStorage>>,
+    req: LookupRequest,
+) -> Result<LookupResponse, Status> {
+    tracing::debug!(
+        "Lookup request received: keys={:?} read_options={:?}",
+        req.keys,
+        req.read_options
+    );
+    let storage = storage.read().await;
+    let mut found = Vec::new();
+    let mut missing = Vec::new();
+
+    for key in &req.keys {
+        let result_entity = storage.get_entity(key);
+        if let Some(entity) = result_entity {
+            found.push(EntityResult {
+                entity: Some(entity.entity.clone()),
+                create_time: Some(entity.create_time.clone()),
+                update_time: Some(entity.update_time.clone()),
+                cursor: vec![],
+                version: entity.version as i64,
+            });
+        } else {
+            let entity = Entity {
+                key: Some(key.clone()),
+                properties: HashMap::new(),
+            };
+            missing.push(EntityResult {
+                entity: Some(entity),
+                create_time: None,
+                update_time: None,
+                cursor: vec![],
+                version: 0,
+            });
+        }
+    }
+    let read_time = system_time_to_timestamp(SystemTime::now());
+
+    Ok(LookupResponse {
+        found,
+        missing,
+        deferred: vec![],
+        transaction: Vec::new(),
+        read_time: Some(read_time),
+    })
+}
